implemented automaton state, changed related files
